move visgeometry-associated files inside of visgeometry
