+++ src/simularium/VisGeometry/types.ts
@@ -1,5 +1,5 @@
 import { Object3D } from "three";
-import PDBModel from "../PDBModel";
+import PDBModel from "./PDBModel";
 import { InstancedMesh } from "./rendering/InstancedMesh";
 
 export interface MeshLoadRequest {
